clevis pin: fix poll bugs
If the Tang server hostname resolves to multiple addresses but is
not accessible at the first address, the request procedure does not
read the correct file descriptor, due to a precedence error when
examining the revents member.

Furthermore, the return value of poll(2) is the number of file
descriptors with events, but iteration of the poll array is
constrained to j < r.  If one event occurred (r = 1), on the second
fd (j = 1), it cannot be reached.

This commit fixes these issues.