Update client-side advertisement validation
First, we ensure that the advertisement is signed by all the
advertised signing keys. Second, in the case where trusted keys
are provided, we ensure that the advertisement is signed by at
least one of the trusted signing keys. Finally, we ensure that
there is at least one recovery key and at least one signing key
in the advertisement.