Run udisks2 as a userspace setuid helper
UDisks2 has an unfortunate property that it keeps internal state of who
called the Unlock() method. Only users who call Unlock() receive
permissions to actually mount the filesystem. This means that calling
Unlock() as root on behalf of the user means the user cannot mount
the resulting unlocked block device.

In order to solve this we make a setuid userpace daemon. The only part
of this binary that runs as root is reading the Tang metadata from the
LUKS header. This metadata is passed to the unprivileged child which
performs all network access and calls Unlock().