Reformulate key management metadata
Don't use extended attributes. They don't work on some filesystems.
Instead, the key use is the file extension and the advertisement
is controlled using a leading '.': if the file is hidden in the
filesystem, it isn't advertised to clients. Futher, the filenames
are now based on timestamps rather than random data.

This allows key management to be done with simple and obvious
file management techniques. Thus, we can retire tang-key-mod.

Additionally, tang-key-groups is now subsumed inside tang-key-gen.