Move database handling out of process
José now ships all the shell commands we need to generate a database
with a shell script. So we setup a systemd path watch on the JWK
directory and, when any changes are detected, we regenerate all the
signatures and create symlinks for the keys based on their thumbprints.
This allows us to move most of the code out of the daemon (and C).

In this new structure, tangd just hands off preconfigured advertisement
files. It is also able to look up JWKs using the thumbprint symlinks
for recovery requests.