Add functions for key manipulation
We currently rely on the tangd-update script to read the keys and
generate signed advertisements as well as JWKs for key derivation.

Whenever there is a change in the directory containing the actual
keys, we run tangd-update through a systemd file watching mechanism,
so that we can have a cache directory with updated advertisements +
JWKs.

As reported in #23 and #24, this mechanism can be unreliable in
certain situations, and having up-to-date information on the keys that
are available is critical to tang, so the idea here is to remove this
dependency on external scripts (e.g. tangd-update) and move this
computation to tang itself.

In this commit we add the related functions for key manipulation so
that in a next step we can start using it in tang.