Allow building on FreeBSD variants (#60)
These changes allow building the project on FreeBSD and it's variants.

The README is changed to note describe the build. A number of 
changes are to tests to remove bash dependencies since bash is
not guaranteed to be installed on FreeBSD.

Additionally while researching the openwrt changes (which are similar 
in many ways) I discovered that the openwrt project is using a somewhat
different configuration. And in fact tang will not run on openwrt as
configured here before. So I updated that configuration to more closely
match the openwrt one.