Fix generation of new keys when no keys are available
When no keys are available, tang creates a new pair of keys, however
currently it checks the total number of keys, including rotated keys,
to decide whether to create new keys.

So not to have issues when all the keys have been rotated, let's check
instead the total number of "regular" keys, the ones that will be
advertised, and if there are none, then tang can create new keys.

This fixes an issue when we do have all keys rotated.
Tests added as well.