tangd.socket: Set ownership and mode of jwkdir
On OSTree-enabled systems (such as Fedora CoreOS), every update
recreates the passwd and group files, which means that the
dynamically-assigned ID and GID of the tang daemon may change,
depending on the order of createuser calls. This patch ensures
that starting the socket unit will fix any ownership issues in
the jwkdir.

This is essentially moving the Fedora %post script[1] out of RPM
and into the systemd unit instead.

[1] https://src.fedoraproject.org/rpms/tang/blob/ba7119863624a55ac4bf79e19acd2a26350896e0/f/tang.spec#_68

Signed-off-by: Stephen Gallagher <sgallagh@redhat.com>